Added user authorization

+++ server/internal/repository/document_repo.go
@@ -23,21 +23,20 @@ func NewDocumentRepository(db *pgxpool.Pool) *DocumentRepository {
 
 func (r *DocumentRepository) Create(doc *models.Document) (int, error) {
 	query := `
-		INSERT INTO documents (title, authors, year, category, file_path, file_size)
-		VALUES ($1, $2, $3, $4, $5, $6)
-		RETURNING id, created_at
-	`
+        INSERT INTO documents (title, authors, year, category, file_path, file_size, user_id)
+        VALUES ($1, $2, $3, $4, $5, $6, $7)
+        RETURNING id, created_at
+    `
 	err := r.db.QueryRow(r.ctx, query,
-		doc.Title, doc.Authors, doc.Year, doc.Category, doc.FilePath, doc.FileSize,
+		doc.Title, doc.Authors, doc.Year, doc.Category, doc.FilePath, doc.FileSize, doc.UserID,
 	).Scan(&doc.ID, &doc.CreatedAt)
 	if err != nil {
 		return 0, fmt.Errorf("insert document: %w", err)
 	}
-
 	return doc.ID, nil
 }
 
-func (r *DocumentRepository) GetByID(id int) (*models.Document, error) {
+func (r *DocumentRepository) GetByID(id, userID int) (*models.Document, error) {
 	query := `
 		SELECT
 			id,
@@ -48,10 +47,10 @@ func (r *DocumentRepository) GetByID(id int) (*models.Document, error) {
 			file_path,
 			file_size,
 			created_at
-		FROM documents WHERE id = $1
+		FROM documents WHERE id = $1 AND user_id = $2
 	`
 	var doc models.Document
-	err := r.db.QueryRow(r.ctx, query, id).Scan(
+	err := r.db.QueryRow(r.ctx, query, id, userID).Scan(
 		&doc.ID, &doc.Title, &doc.Authors, &doc.Year, &doc.Category,
 		&doc.FilePath, &doc.FileSize, &doc.CreatedAt,
 	)
@@ -62,7 +61,7 @@ func (r *DocumentRepository) GetByID(id int) (*models.Document, error) {
 }
 
 func (r *DocumentRepository) GetAll(
-	limit, offset int,
+	userID, limit, offset int,
 ) ([]models.Document, error) {
 	if limit <= 0 {
 		limit = 20
@@ -77,10 +76,10 @@ func (r *DocumentRepository) GetAll(
 			file_path,
 			file_size,
 			created_at
-        FROM documents ORDER BY created_at DESC
-        LIMIT $1 OFFSET $2
+		FROM documents WHERE user_id = $1 ORDER BY created_at DESC
+        LIMIT $2 OFFSET $3
     `
-	rows, err := r.db.Query(r.ctx, query, limit, offset)
+	rows, err := r.db.Query(r.ctx, query, userID, limit, offset)
 	if err != nil {
 		return nil, fmt.Errorf("get all documents: %w", err)
 	}
@@ -100,9 +99,9 @@ func (r *DocumentRepository) GetAll(
 	return docs, nil
 }
 
-func (r *DocumentRepository) Delete(id int) error {
-	query := `DELETE FROM documents WHERE id = $1`
-	cmdTag, err := r.db.Exec(r.ctx, query, id)
+func (r *DocumentRepository) Delete(id, userID int) error {
+	query := `DELETE FROM documents WHERE id = $1 AND user_id = $2`
+	cmdTag, err := r.db.Exec(r.ctx, query, id, userID)
 	if err != nil {
 		return fmt.Errorf("delete document: %w", err)
 	}
